+++ analog.h
@@ -11,6 +11,10 @@
  * If that was necessary, using keyboard keys to rotate the control stick
  * would have been undetected by the game for such a precision requirement.
  *
+ * Negative values should, in theory, yield the same results of spins done on
+ * the absolute value of the negative precision clockwise instead of counter-
+ * clockwise, but this has not been tested to work as well as CCW spins.
+ *
  * Results from the Mecha Fly-Guy stick test in Mario Party:
  *     f(60) =  10 spins / 10 seconds (barely enough to draw Paddle Battle)
  *     f(24) =  25 spins / 10 seconds (almost enough to win Power Pedal)
@@ -27,6 +31,6 @@ extern long arcs_per_second;
 
 extern signed char clamp_stick(signed long magnitude);
 extern int stick_range(void);
-extern void stick_rotate(signed char * x, signed char * y, double degrees);
+extern void stick_rotate(signed char * x, signed char * y, float degrees);
 
 #endif
